requirements.txt
python-nmap~=0.7.1
click~=8.1.7
requests~=2.31.0